Remade the memory killswitch, moved compare funtion into the header file and over small changes
